requirements-cli.txt
# Minimal dependencies for CLI version only
openai>=1.0.0
python-dotenv>=1.0.0
